% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areacircle.R
\name{area_of_a_circle}
\alias{area_of_a_circle}
\title{Calculate the Area of a Circle}
\usage{
area_of_a_circle(r)
}
\arguments{
\item{r}{The radius of the circle}
}
\value{
The area of the circle
}
\description{
To calculate the area of a circle with the given radius.
}
\examples{
area_of_a_circle(2)
\dontrun{area_of_a_circle(2)}
}
